% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droputils.R
\name{print.dropit}
\alias{print.dropit}
\title{Print method for dropit objects}
\usage{
\method{print}{dropit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{dropit}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print method for dropit objects
}
