% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droputils.R
\name{print.dropit_log}
\alias{print.dropit_log}
\title{Print method for dropit_log objects}
\usage{
\method{print}{dropit_log}(x, ...)
}
\arguments{
\item{x}{An object of class \code{dropit_log}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print method for dropit_log objects
}
